+++ chapter 2/3-final_app/food_fact_functions/open_state_selector_callback.py
@@ -1,0 +1,2 @@
+def open_state_selector(state):
+    state.open_states = True
