+++ chapter 2/3-final_app/pages/food_facts.py
@@ -3,9 +3,11 @@ import plotly.express as px
 import taipy.gui.builder as tgb
 import us
 from food_fact_functions.add_row_callback import add_row
+from food_fact_functions.charts import create_fig_states
 from food_fact_functions.delete_row_callback import delete_row
+from food_fact_functions.edit_note_callback import edit_note
 from food_fact_functions.initiate_sales import clean_sales_data, update_df_sales
-from sqlalchemy import true
+from food_fact_functions.open_state_selector_callback import open_state_selector
 
 ###########################
 ## Initial values        ##
@@ -34,32 +36,12 @@ lov_metrics = ["FAH", "FAFH", "Total"]
 metric = "Total"
 
 
-def create_fig_states(df_sales, metric):
-    fig_states = px.choropleth(
-        df_sales,
-        locations="State_Code",
-        locationmode="USA-states",
-        color=metric,
-        title=f"Values per state for {metric} sales",
-        scope="usa",
-        color_continuous_scale=px.colors.sequential.ice_r,
-    )
-    return fig_states
-
-
 fig_states = create_fig_states(df_sales, "Total")
 
 
 ###########################
-## Callbacks             ##
+## Main Callback         ##
 ###########################
-def edit_note(state, var_name, payload):
-    if payload["col"] == "Note":
-        df_sales_new = state.df_sales.copy()
-        df_sales_new.loc[payload["index"], payload["col"]] = payload["value"]
-        state.df_sales = df_sales_new
-    else:
-        None
 
 
 def update_sales(state, var_name, payload):
@@ -69,8 +51,6 @@ def update_sales(state, var_name, payload):
     df_sales_copy.reset_index(drop=True, inplace=True)
 
     filter_condition = pd.Series([True] * len(df_sales_copy))
-    print(df_sales_copy)
-    print(filter_condition)
     if state.selected_year != "All":
         filter_condition &= df_sales_copy["Year"] == state.selected_year
 
@@ -86,28 +66,13 @@ def update_sales(state, var_name, payload):
     state.fig_states = create_fig_states(state.df_sales, state.metric)
 
 
-def open_state_selector(state):
-    state.open_states = True
-
-
 ###########################
-## Page             ##
+## Page                 ##
 ###########################
 
 
 with tgb.Page() as food_fact_page:
     tgb.text("# Food facts 📊", mode="md", class_name="color-secondary header")
-    tgb.table(
-        data="{df_sales}",
-        height="60vh",
-        filter=True,
-        hover_text="USDA Data ",
-        on_edit=edit_note,
-        on_add=add_row,
-        on_delete=delete_row,
-        class_name="p0 m0",
-        nan_value=0,
-    )
 
     with tgb.pane(open="{open_states}"):
         tgb.text("## Select states", mode="md", class_name="color-secondary")
@@ -117,7 +82,7 @@ with tgb.Page() as food_fact_page:
             on_change=update_sales,
             label="select states",
             multiple=True,
-            mode="checkbox",
+            # mode="checkbox",
         )
 
     with tgb.layout("1 1 1 1"):
@@ -165,3 +130,15 @@ with tgb.Page() as food_fact_page:
             figure="{fig_states}",
             class_name="p0 m0",
         )
+
+    tgb.table(
+        data="{df_sales}",
+        height="60vh",
+        filter=True,
+        hover_text="USDA Data ",
+        on_edit=edit_note,
+        on_add=add_row,
+        on_delete=delete_row,
+        class_name="p0 m0",
+        nan_value=0,
+    )
